modifiche sostanziali fatte in gruppo per integrare funzionalita database

+++ app/src/main/java/it/unicam/cs/FilieraAgricola/Product/SingleProduct.java
@@ -3,7 +3,7 @@ package it.unicam.cs.FilieraAgricola.Product;
 
 public class SingleProduct extends Product{
 
-    public SingleProduct(int productID, String productName, String productDescription, double productPrice, int productQuantity, ProductState ProductState) {
+    public SingleProduct(long productID, String productName, String productDescription, double productPrice, int productQuantity, ProductState ProductState) {
         super(productID, productName, productDescription, productPrice, productQuantity, ProductState);
     }
